+++ client/src/app/App.jsx
@@ -1,13 +1,22 @@
-import { BrowserRouter, Navigate, Route, Routes } from "react-router";
-import LoginPage from "./../features/login/LoginPage.jsx";
-import RecordsPage from "../features/records/RecordsPage.jsx";
-import RegisterPage from "../features/login/RegisterPage.jsx";
-import StampPage from "../features/records/stamp/StampPage.jsx";
-import MapPage from "../features/records/map/MapPage.jsx";
+import { BrowserRouter, Navigate, Route, Routes } from 'react-router';
+import LoginPage from './../features/login/LoginPage.jsx';
+import RecordsPage from '../features/records/RecordsPage.jsx';
+import RegisterPage from '../features/login/RegisterPage.jsx';
+import StampPage from '../features/records/stamp/StampPage.jsx';
+import MapPage from '../features/records/map/MapPage.jsx';
+import React, { useState, createContext } from 'react';
 
-const App = ()=>{
-    return (
-        <BrowserRouter>
+export const context = createContext();
+
+const App = () => {
+  const [postRendering, setPostRendering] = useState(false);
+
+  function rendering() {
+    setPostRendering(!postRendering);
+  }
+  return (
+    <context.Provider value={{ rendering, postRendering }}>
+      <BrowserRouter>
         <Routes>
           <Route path="/" element={<Navigate to="/login" />} />
           <Route path="/login" element={<LoginPage />} />
@@ -17,7 +26,8 @@ const App = ()=>{
           <Route path="/stamp" element={<StampPage />} />
         </Routes>
       </BrowserRouter>
-    );
-}
+    </context.Provider>
+  );
+};
 
-export default App; 
+export default App;
